comment id is a bigint

+++ database/migrations/2018_12_13_121903_add_reportable_type_to_user_reports.php
@@ -15,7 +15,7 @@ class AddReportableTypeToUserReports extends Migration
     {
         Schema::table('osu_user_reports', function (Blueprint $table) {
             $table->enum('reportable_type', ['user', 'score', 'comment'])->nullable()->after('user_id');
-            $table->unsignedInteger('reportable_id')->nullable()->after('mode');
+            $table->unsignedBigInteger('reportable_id')->nullable()->after('mode');
             $table->unique(['reporter_id', 'user_id', 'reportable_type', 'mode', 'reportable_id'], 'unique-reportable');
             $table->index(['reportable_type', 'reportable_id'], 'reportable');
             $table->index(['score_id', 'mode'], 'score');
